Support pacing commands/logins
When we introduced limiting the max number of outstanding commands to a
AAA server, we didn't pace the commands. As a result, in a network with
fast device responses, we could still overwhelm the AAA server.
Furthermore, some networks only contact the AAA server for
authentication (at login time) while others also contact it for
authorization (per command). Finally, even with the best of
implementations, the AAA server can sometimes fail to
authenticate/authorize even though the credentials are valid. To handle
all these different cases, this patch introduces the following behavior
and configuration knobs:

  * Instead of treating the max-cmd-pipeline parameter as merely
    controlling the max outstanding requests, we instead use this as the
    rate/sec. Thus, a value of 10 is treated as 10 requests/sec which
    then gets converted into a 100ms sleep between requests, thus evenly
    pacing out the 10 requests across the second.
  * We add a per-cmd-auth parameter to the devices section of the
    inventory and use that to control whether we pace commands or only
    logins.
  * We introduce another parameter, retries-on-auth-fail, in the devices
    section of the inventory, and use this to control how many times we
    retry if an authentication fails. This value is reset on successful
    authentication. By default, we don't retry.

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>